Move from float to double

+++ src/App/Annotation.cpp
@@ -36,7 +36,7 @@ PROPERTY_SOURCE(App::Annotation, App::DocumentObject)
 Annotation::Annotation() 
 {
     ADD_PROPERTY(LabelText ,(""));
-    ADD_PROPERTY(Position,(Base::Vector3f()));
+    ADD_PROPERTY(Position,(Base::Vector3d()));
 }
 
 Annotation::~Annotation()
@@ -51,8 +51,8 @@ PROPERTY_SOURCE(App::AnnotationLabel, App::DocumentObject)
 AnnotationLabel::AnnotationLabel() 
 {
     ADD_PROPERTY_TYPE(LabelText,(""),"Label",Prop_Output,"Text label of the annotation");
-    ADD_PROPERTY_TYPE(BasePosition,(Base::Vector3f()),"Label",Prop_Output,"Base position");
-    ADD_PROPERTY_TYPE(TextPosition,(Base::Vector3f()),"Label",Prop_Output,"Text position");
+    ADD_PROPERTY_TYPE(BasePosition,(Base::Vector3d()),"Label",Prop_Output,"Base position");
+    ADD_PROPERTY_TYPE(TextPosition,(Base::Vector3d()),"Label",Prop_Output,"Text position");
 }
 
 AnnotationLabel::~AnnotationLabel()
